pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "baseline"
version = "1.0.0"
description = ""
license = "Apache-2.0"
dependencies = [
    "flwr[simulation]>=1.15.2",
    "flwr-datasets[vision]>=0.5.0",
    "torch==2.5.1",
    "torchvision==0.20.1",
    "wandb",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.flwr.app]
publisher = "julioman"

[tool.flwr.app.components]
serverapp = "baseline.server_app:app"
clientapp = "baseline.client_app:app"

[tool.flwr.app.config]
num-server-rounds = 3
fraction-fit = 0.5
local-epochs = 1

[tool.flwr.federations]
default = "local-simulation"

[tool.flwr.federations.local-simulation]
options.num-supernodes = 10

[tool.flwr.federations.larger-sim]
options.num-supernodes=100
options.backend.client-resources.num-cpus = 1

[tool.flwr.federations.gpu-sim]
options.num-supernodes=1000
options.backend.client-resources.num-cpus = 1
options.backend.client-resources.num-gpus = 0.16